pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "dont_forget"
version = "0.0.1"
dependencies = [
    "annotated-types==0.7.0",
    "asgiref==3.8.1",
    "cffi==1.16.0",
    "contextlib2==21.6.0",
    "cryptography==43.0.0",
    "diff-match-patch==20230430",
    "Django==5.0.7",
    "django-cors-headers==4.4.0",
    "django-environ==0.11.2",
    "django-extensions==3.2.3",
    "django-import-export==4.1.1",
    "django-ninja==1.2.2",
    "django-ninja-crud==0.6.0",
    "django-ninja-extra==0.21.2",
    "django-ninja-jwt==5.3.2",
    "django-simple-history==3.7.0",
    "injector==0.22.0",
    "pillow==10.4.0",
    "psycopg==3.2.1",
    "psycopg-c==3.2.1",
    "pycparser==2.22",
    "pydantic==2.8.2",
    "pydantic_core==2.20.1",
    "PyJWT==2.9.0",
    "sqlparse==0.5.1",
    "tablib==3.5.0",
    "typing_extensions==4.12.2",
]

[project.optional-dependencies]
dev = [
  "black",
  "flake8",
  "Flake8-pyproject",
  "isort"
]

[tool.isort]
profile = "black"
line_length = 125

[tool.black]
line-length = 125
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | migrations
)/
'''

[tool.flake8]
ignore = ['E231', 'E241']
per-file-ignores = [
    '__init__.py:F401',
]
max-line-length = 125
count = true